Update RandomGen casting for clang

+++ new_src/utils/randomgen.h
@@ -52,7 +52,7 @@ class RandomGen {
       for (uint64_t i = 0; i < v.size(); ++i) {
         this->seed = Utils::sha3(this->seed.get());
         //std::cout << this->seed.hex() << std::endl; // Uncomment to print seed
-        uint64_t n = i + this->seed.toUint256() % (v.size() - i);
+        uint64_t n = uint64_t(i + this->seed.toUint256() % (v.size() - i));
         std::swap(v[n], v[i]);
       }
       this->seedLock.unlock();
